feat: re-organize code to get ready to support multiple DB types
