+++ pkg/plugin/postgresql/constants.go
@@ -1,4 +1,4 @@
-package plugin
+package postgresql
 
 import (
 	core "k8s.io/api/core/v1"
