+++ pkg/plugin/postgresql/kubedb_postgres.go
@@ -1,4 +1,4 @@
-package plugin
+package postgresql
 
 import (
 	"fmt"
